pyproject.toml
[tool.poetry]
name = "training"
version = "0.1.0"
description = "FÃ¡brica de Software training"
authors = ["Luis <devcorujam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^5.0.0"}

[tool.poetry.dev-dependencies]
model-bakery = "^1.2.1"
flake8 = "^3.9.0"
black = "^20.8b1"
isort = "^5.8.0"
pre-commit = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
